requirements.txt
streamlit
pdfplumber
PyPDF2
pillow
